Arm backend: Annotate ADD/SUB with indepenedent observers
We were annotating the ADD/SUB with a shared observer resulting in
the same quantisation parameters on the two inputs even if
we were adding  numbers in different ranges(positive tensor to a tensor
with positive and negative values). As a result,
the quantisation parameters were suboptimal. This change annotates
the operator with independent observers and changes how we rescale
the two inputs to bring them to the same range. Added a unit test
of a resnet model. Lowered the number of channels on a few unit tests
in order to keep the Total SRAM Used below 2MB for the Ethos-U55 to fit
within the memory limit of the Corstone-300.

Change-Id: I7adde636f901c9df6b779d946a157e66fd12e24e